+++ docs/.vuepress/config.js
@@ -3,7 +3,7 @@ const path = require("path");
 module.exports = context => ({
   base: process.env.VUEPRESS_BASE || "/",
   title: "NoneBot",
-  description: "基于 酷Q 的 Python 异步 QQ 机器人框架",
+  description: "跨平台 Python 异步 QQ 机器人框架",
   markdown: {
     lineNumbers: true
   },
@@ -56,7 +56,7 @@ module.exports = context => ({
     "/": {
       lang: "zh-CN",
       title: "NoneBot",
-      description: "基于 酷Q 的 Python 异步 QQ 机器人框架"
+      description: "跨平台 Python 异步 QQ 机器人框架"
     }
   },
 
@@ -116,7 +116,7 @@ module.exports = context => ({
               title: "协议适配",
               collapsable: false,
               sidebar: "auto",
-              children: ["cqhttp-guide", "ding-guide"]
+              children: ["cqhttp-guide", "ding-guide", "mirai-guide"]
             }
           ],
           "/advanced/": [
@@ -209,6 +209,10 @@ module.exports = context => ({
                 {
                   title: "nonebot.adapters.ding 模块",
                   path: "adapters/ding"
+                },
+                {
+                  title: "nonebot.adapters.mirai 模块",
+                  path: "adapters/mirai"
                 }
               ]
             }