+++ nonebot/plugin.py
@@ -19,7 +19,7 @@ from nonebot.log import logger
 from nonebot.matcher import Matcher
 from nonebot.permission import Permission
 from nonebot.typing import T_State, T_StateFactory, T_Handler, T_RuleChecker
-from nonebot.rule import Rule, startswith, endswith, keyword, command, regex
+from nonebot.rule import Rule, startswith, endswith, keyword, command, shell_command, ArgumentParser, regex
 
 if TYPE_CHECKING:
     from nonebot.adapters import Bot, Event
@@ -425,7 +425,7 @@ def on_command(cmd: Union[str, Tuple[str, ...]],
         segment = message.pop(0)
         new_message = message.__class__(
             str(segment)
-            [len(state["_prefix"]["raw_command"]):].strip())  # type: ignore
+            [len(state["_prefix"]["raw_command"]):].lstrip())  # type: ignore
         for new_segment in reversed(new_message):
             message.insert(0, new_segment)
 
@@ -434,6 +434,57 @@ def on_command(cmd: Union[str, Tuple[str, ...]],
 
     commands = set([cmd]) | (aliases or set())
     return on_message(command(*commands) & rule, handlers=handlers, **kwargs)
+
+
+def on_shell_command(cmd: Union[str, Tuple[str, ...]],
+                     rule: Optional[Union[Rule, T_RuleChecker]] = None,
+                     aliases: Optional[Set[Union[str, Tuple[str, ...]]]] = None,
+                     parser: Optional[ArgumentParser] = None,
+                     **kwargs) -> Type[Matcher]:
+    """
+    :说明:
+
+      注册一个支持 ``shell_like`` 解析参数的命令消息事件响应器。
+
+      与普通的 ``on_command`` 不同的是，在添加 ``parser`` 参数时, 响应器会自动处理消息。
+
+      并将用户输入的原始参数列表保存在 ``state["argv"]``, ``parser`` 处理的参数保存在 ``state["args"]`` 中
+
+    :参数:
+
+      * ``cmd: Union[str, Tuple[str, ...]]``: 指定命令内容
+      * ``rule: Optional[Union[Rule, T_RuleChecker]]``: 事件响应规则
+      * ``aliases: Optional[Set[Union[str, Tuple[str, ...]]]]``: 命令别名
+      * ``parser: Optional[ArgumentParser]``: ``nonebot.rule.ArgumentParser`` 对象
+      * ``permission: Optional[Permission]``: 事件响应权限
+      * ``handlers: Optional[List[T_Handler]]``: 事件处理函数列表
+      * ``temp: bool``: 是否为临时事件响应器（仅执行一次）
+      * ``priority: int``: 事件响应器优先级
+      * ``block: bool``: 是否阻止事件向更低优先级传递
+      * ``state: Optional[T_State]``: 默认 state
+      * ``state_factory: Optional[T_StateFactory]``: 默认 state 的工厂函数
+
+    :返回:
+
+      - ``Type[Matcher]``
+    """
+
+    async def _strip_cmd(bot: "Bot", event: "Event", state: T_State):
+        message = event.get_message()
+        segment = message.pop(0)
+        new_message = message.__class__(
+            str(segment)
+            [len(state["_prefix"]["raw_command"]):].strip())  # type: ignore
+        for new_segment in reversed(new_message):
+            message.insert(0, new_segment)
+
+    handlers = kwargs.pop("handlers", [])
+    handlers.insert(0, _strip_cmd)
+
+    commands = set([cmd]) | (aliases or set())
+    return on_message(shell_command(*commands, parser=parser) & rule,
+                      handlers=handlers,
+                      **kwargs)
 
 
 def on_regex(pattern: str,
@@ -512,6 +563,29 @@ class CommandGroup:
         final_kwargs = self.base_kwargs.copy()
         final_kwargs.update(kwargs)
         return on_command(cmd, **final_kwargs)
+
+    def shell_command(self, cmd: Union[str, Tuple[str, ...]],
+                      **kwargs) -> Type[Matcher]:
+        """
+        :说明:
+
+          注册一个新的命令。
+
+        :参数:
+
+          * ``cmd: Union[str, Tuple[str, ...]]``: 命令前缀
+          * ``**kwargs``: 其他传递给 ``on_command`` 的参数，将会覆盖命令组默认值
+
+        :返回:
+
+          - ``Type[Matcher]``
+        """
+        sub_cmd = (cmd,) if isinstance(cmd, str) else cmd
+        cmd = self.basecmd + sub_cmd
+
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        return on_shell_command(cmd, **final_kwargs)
 
 
 class MatcherGroup:
@@ -671,11 +745,7 @@ class MatcherGroup:
         self.matchers.append(matcher)
         return matcher
 
-    def on_startswith(self,
-                      msg: str,
-                      rule: Optional[Optional[Union[Rule,
-                                                    T_RuleChecker]]] = None,
-                      **kwargs) -> Type[Matcher]:
+    def on_startswith(self, msg: str, **kwargs) -> Type[Matcher]:
         """
         :说明:
 
@@ -697,12 +767,14 @@ class MatcherGroup:
 
           - ``Type[Matcher]``
         """
-        return self.on_message(rule=startswith(msg) & rule, **kwargs)
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_startswith(msg, **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
-    def on_endswith(self,
-                    msg: str,
-                    rule: Optional[Optional[Union[Rule, T_RuleChecker]]] = None,
-                    **kwargs) -> Type[Matcher]:
+    def on_endswith(self, msg: str, **kwargs) -> Type[Matcher]:
         """
         :说明:
 
@@ -724,12 +796,14 @@ class MatcherGroup:
 
           - ``Type[Matcher]``
         """
-        return self.on_message(rule=endswith(msg) & rule, **kwargs)
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_endswith(msg, **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
-    def on_keyword(self,
-                   keywords: Set[str],
-                   rule: Optional[Union[Rule, T_RuleChecker]] = None,
-                   **kwargs) -> Type[Matcher]:
+    def on_keyword(self, keywords: Set[str], **kwargs) -> Type[Matcher]:
         """
         :说明:
 
@@ -751,11 +825,15 @@ class MatcherGroup:
 
           - ``Type[Matcher]``
         """
-        return self.on_message(rule=keyword(*keywords) & rule, **kwargs)
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_keyword(keywords, **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
     def on_command(self,
                    cmd: Union[str, Tuple[str, ...]],
-                   rule: Optional[Union[Rule, T_RuleChecker]] = None,
                    aliases: Optional[Set[Union[str, Tuple[str, ...]]]] = None,
                    **kwargs) -> Type[Matcher]:
         """
@@ -768,8 +846,8 @@ class MatcherGroup:
         :参数:
 
           * ``cmd: Union[str, Tuple[str, ...]]``: 指定命令内容
-          * ``rule: Optional[Union[Rule, T_RuleChecker]]``: 事件响应规则
           * ``aliases: Optional[Set[Union[str, Tuple[str, ...]]]]``: 命令别名
+          * ``rule: Optional[Union[Rule, T_RuleChecker]]``: 事件响应规则
           * ``permission: Optional[Permission]``: 事件响应权限
           * ``handlers: Optional[List[T_Handler]]``: 事件处理函数列表
           * ``temp: bool``: 是否为临时事件响应器（仅执行一次）
@@ -782,28 +860,59 @@ class MatcherGroup:
 
           - ``Type[Matcher]``
         """
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_command(cmd, aliases=aliases, **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
-        async def _strip_cmd(bot: "Bot", event: "Event", state: T_State):
-            message = event.get_message()
-            segment = message.pop(0)
-            new_message = message.__class__(
-                str(segment)
-                [len(state["_prefix"]["raw_command"]):].strip())  # type: ignore
-            for new_segment in reversed(new_message):
-                message.insert(0, new_segment)
+    def on_shell_command(self,
+                         cmd: Union[str, Tuple[str, ...]],
+                         aliases: Optional[Set[Union[str, Tuple[str,
+                                                                ...]]]] = None,
+                         parser: Optional[ArgumentParser] = None,
+                         **kwargs) -> Type[Matcher]:
+        """
+        :说明:
 
-        handlers = kwargs.pop("handlers", [])
-        handlers.insert(0, _strip_cmd)
+          注册一个支持 ``shell_like`` 解析参数的命令消息事件响应器。
 
-        commands = set([cmd]) | (aliases or set())
-        return self.on_message(rule=command(*commands) & rule,
-                               handlers=handlers,
-                               **kwargs)
+          与普通的 ``on_command`` 不同的是，在添加 ``parser`` 参数时, 响应器会自动处理消息。
+
+          并将用户输入的原始参数列表保存在 ``state["argv"]``, ``parser`` 处理的参数保存在 ``state["args"]`` 中
+
+        :参数:
+
+          * ``cmd: Union[str, Tuple[str, ...]]``: 指定命令内容
+          * ``aliases: Optional[Set[Union[str, Tuple[str, ...]]]]``: 命令别名
+          * ``parser: Optional[ArgumentParser]``: ``nonebot.rule.ArgumentParser`` 对象
+          * ``rule: Optional[Union[Rule, T_RuleChecker]]``: 事件响应规则
+          * ``permission: Optional[Permission]``: 事件响应权限
+          * ``handlers: Optional[List[T_Handler]]``: 事件处理函数列表
+          * ``temp: bool``: 是否为临时事件响应器（仅执行一次）
+          * ``priority: int``: 事件响应器优先级
+          * ``block: bool``: 是否阻止事件向更低优先级传递
+          * ``state: Optional[T_State]``: 默认 state
+          * ``state_factory: Optional[T_StateFactory]``: 默认 state 的工厂函数
+
+        :返回:
+
+          - ``Type[Matcher]``
+        """
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_shell_command(cmd,
+                                   aliases=aliases,
+                                   parser=parser,
+                                   **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
     def on_regex(self,
                  pattern: str,
                  flags: Union[int, re.RegexFlag] = 0,
-                 rule: Optional[Rule] = None,
                  **kwargs) -> Type[Matcher]:
         """
         :说明:
@@ -829,7 +938,12 @@ class MatcherGroup:
 
           - ``Type[Matcher]``
         """
-        return self.on_message(rule=regex(pattern, flags) & rule, **kwargs)
+        final_kwargs = self.base_kwargs.copy()
+        final_kwargs.update(kwargs)
+        final_kwargs.pop("type", None)
+        matcher = on_regex(pattern, flags=flags, **final_kwargs)
+        self.matchers.append(matcher)
+        return matcher
 
 
 def load_plugin(module_path: str) -> Optional[Plugin]:
@@ -933,7 +1047,7 @@ def load_plugins(*plugin_dir: str) -> Set[Plugin]:
     return loaded_plugins
 
 
-def load_builtin_plugins() -> Optional[Plugin]:
+def load_builtin_plugins(name: str = "echo") -> Optional[Plugin]:
     """
     :说明:
 
@@ -943,7 +1057,7 @@ def load_builtin_plugins() -> Optional[Plugin]:
 
       - ``Plugin``
     """
-    return load_plugin("nonebot.plugins.base")
+    return load_plugin(f"nonebot.plugins.{name}")
 
 
 def get_plugin(name: str) -> Optional[Plugin]: